backend/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.26.0
spotipy==2.19.0
